Moved the license_feature_entitlement and license_software_entitlement to api namespace and renamed the error class for software and feature entitlment
