+++ components/ruby/lib/chef_licensing.rb
@@ -1,17 +1,17 @@
 require "chef_licensing/version"
-require "chef_licensing/license_feature_entitlement"
+require "chef_licensing/api/license_feature_entitlement"
 require "chef_licensing/license_key_fetcher"
 require "chef_licensing/config"
-require "chef_licensing/license_software_entitlement"
+require "chef_licensing/api/license_software_entitlement"
 
 module ChefLicensing
   class << self
     def check_feature_entitlement!(feature_name: nil, feature_id: nil)
-      ChefLicensing::LicenseFeatureEntitlement.check_entitlement!(license_keys: license_keys, feature_name: feature_name)
+      ChefLicensing::Api::LicenseFeatureEntitlement.check_entitlement!(license_keys: license_keys, feature_name: feature_name, feature_id: feature_id)
     end
 
     def check_software_entitlement!(software_entitlement_name: nil, software_entitlement_id: nil)
-      ChefLicensing::LicenseSoftwareEntitlement.check!(license_keys: license_keys, software_entitlement_name: software_entitlement_name, software_entitlement_id: software_entitlement_id)
+      ChefLicensing::Api::LicenseSoftwareEntitlement.check!(license_keys: license_keys, software_entitlement_name: software_entitlement_name, software_entitlement_id: software_entitlement_id)
     end
 
     # @note no in-memory caching of the licenses so that it fetches updated licenses always
