+++ components/ruby/lib/chef_licensing/exceptions/feature_not_entitled.rb
@@ -1,9 +1,9 @@
 require_relative "error"
 
 module ChefLicensing
-  class InvalidEntitlement < Error
+  class FeatureNotEntitled < Error
     def message
-      super || "Invalid Entitlement"
+      super || "Feature not entitled"
     end
   end
 end
