+++ components/ruby/lib/chef_licensing/exceptions/software_not_entitled.rb
@@ -1,0 +1,9 @@
+require_relative "error"
+
+module ChefLicensing
+  class SoftwareNotEntitled < Error
+    def message
+      super || "Software not entitled"
+    end
+  end
+end
